% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{course_slug}
\alias{course_slug}
\title{The course slug used in the cartridge filename}
\usage{
course_slug(course)
}
\arguments{
\item{course}{A parsed \code{course.yml}.}
}
\description{
\verb{slug:} in \code{course.yml} when set, else the slugified Canvas course code,
else the slugified \code{code}. Decision D1 of the portable plan.
}
