% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdiff.R
\name{unpack_cartridge}
\alias{unpack_cartridge}
\title{Unpack a cartridge for inspection}
\usage{
unpack_cartridge(zip)
}
\arguments{
\item{zip}{Path to a Common Cartridge archive.}
}
\value{
The temporary directory containing the unpacked cartridge.
}
\description{
Unpack a cartridge for inspection
}
